Attempt to fix windows tests by adding a windows impl of atomic write.
